[refactoring] test case 분리
