+++ src/test/kotlin/step2/CalculatorTest.kt
@@ -1,8 +1,6 @@
 package step2
 
 import org.assertj.core.api.Assertions.assertThat
-import org.assertj.core.api.Assertions.assertThatThrownBy
-import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
 
@@ -46,33 +44,6 @@ class CalculatorTest {
         assertThat(result).isEqualTo(expected)
     }
 
-    @Test
-    fun null체크() {
-        assertThatThrownBy {
-            Calculator().calculate(null)
-        }.isInstanceOf(IllegalArgumentException::class.java)
-    }
-
-    @ParameterizedTest
-    @CsvSource(
-        value = ["*13+4", "%20+5"],
-    )
-    fun 문자열이_맨앞으로_올_때(input: String) {
-        assertThatThrownBy {
-            Calculator().calculate(input)
-        }.isInstanceOf(IllegalArgumentException::class.java)
-    }
-
-    @ParameterizedTest
-    @CsvSource(
-        value = ["3$4", "6!2"],
-    )
-    fun 사칙연산이_아닐_때(input: String) {
-        assertThatThrownBy {
-            Calculator().calculate(input)
-        }.isInstanceOf(IllegalArgumentException::class.java)
-    }
-
     @ParameterizedTest
     @CsvSource(
         value = ["2+3*4/2=10", "3*4-2*2=20", "0/3*4+3=3"], delimiter = '='
@@ -80,15 +51,5 @@ class CalculatorTest {
     fun 슈퍼_사칙_연산(input: String, expected: String) {
         val result = Calculator().calculate(input)
         assertThat(result).isEqualTo(expected)
-    }
-
-    @ParameterizedTest
-    @CsvSource(
-        value = ["2+3/0*4"]
-    )
-    fun Zero로_나누기(input: String) {
-        assertThatThrownBy {
-            Calculator().calculate(input)
-        }.isInstanceOf(IllegalStateException::class.java)
     }
 }
